# ams/api/openapi.yaml
openapi: 3.1.0
info:
  title: Enterprise Access Management System (AMS) API
  version: 1.0.0
  description: >
    Unified API for visitor management, badge lifecycle, physical access validation,
    approvals, occupancy/evacuation, and audit evidence. Errors follow RFC 7807.
    All POST/PATCH operations require an Idempotency-Key header. Collections use
    cursor pagination (page[size], page[after], page[before]).
  contact:
    name: AMS Platform Team
    email: ams-platform@example.com
  license:
    name: Proprietary
servers:
  - url: https://api.ams.example.com/v1
    description: Production (Front Door -> APIM)
  - url: https://api.uat.ams.example.com/v1
    description: UAT
security:
  - oauth2: []
...
paths:
...
  /badges:
    post:
      tags: [Badges]
      operationId: issueBadge
      summary: Request and issue a badge for a cardholder
      security:
        - oauth2: [ams.badges.write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BadgeIssueRequest' }
      responses:
        '201':
          description: Badge issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Badge' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationProblem' }
    get:
      tags: [Badges]
      operationId: listBadges
      summary: List badges (cursor-paged)
      security:
        - oauth2: [ams.badges.read]
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/PageAfter'
        - $ref: '#/components/parameters/PageBefore'
        - name: filter[cardholderId]
          in: query
          schema: { type: string, format: uuid }
        - name: filter[state]
          in: query
          schema: { type: string, enum: [REQUESTED, ISSUED, ACTIVE, SUSPENDED, REVOKED, EXPIRED, LOST, RETURNED] }
        - name: sort
          in: query
          schema: { type: string, default: "-issuedAt" }
      responses:
        '200':
          description: Page of badges
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BadgePage' }
  /badges/{badgeId}:
    parameters:
      - $ref: '#/components/parameters/BadgeId'
    get:
      tags: [Badges]
      operationId: getBadge
      summary: Badge current state
      security:
        - oauth2: [ams.badges.read]
      responses:
        '200':
          description: Badge
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Badge' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /badges/{badgeId}/revoke:
    parameters:
      - $ref: '#/components/parameters/BadgeId'
    post:
      tags: [Badges]
      operationId: revokeBadge
      summary: Revoke a badge (propagates to edges in <= 5 s P99)
      security:
        - oauth2: [ams.badges.admin]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReasonRequest' }
            examples:
              lost:
                value: { reason: "REPORTED_LOST", comment: "Cardholder reported loss via self-service" }
      responses:
        '200':
          description: Revoked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Badge' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: Microsoft Entra ID (OIDC). Humans use authorization-code + PKCE; services use client credentials.
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/common/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/common/oauth2/v2.0/token
          scopes:
            ams.visits.read: Read visits
            ams.visits.write: Create/amend visits
            ams.visits.checkin: Check visitors in/out
            ams.badges.read: Read badges
            ams.badges.write: Issue/manage badges
            ams.badges.admin: Revoke badges
            ams.validation.execute: Execute badge validations (edge)
            ams.access.request: Request access
            ams.access.read: Read access requests
            ams.access.approve: Approve/reject/delegate
            ams.sites.read: Read sites
            ams.occupancy.read: Read occupancy
            ams.evacuation.activate: Activate evacuations
            ams.evacuation.read: Read evacuation reports
            ams.audit.read: Query audit trail
            ams.webhooks.manage: Manage webhook subscriptions
        clientCredentials:
          tokenUrl: https://login.microsoftonline.com/common/oauth2/v2.0/token
          scopes:
            ams.validation.execute: Execute badge validations (edge)
            ams.webhooks.manage: Manage webhook subscriptions
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: UUID per logical operation. Replays return the original response with Idempotency-Replayed true.
      schema: { type: string, format: uuid }
...
    PageSize:
      name: page[size]
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 25 }
    PageAfter:
      name: page[after]
      in: query
      description: Opaque cursor; return items after this position.
      schema: { type: string }
    PageBefore:
      name: page[before]
      in: query
      description: Opaque cursor; return items before this position.
      schema: { type: string }
...
    BadgeId:
      name: badgeId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    NotFound:
      description: Not found (or hidden by ABAC)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: State conflict (invalid transition or concurrency)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            invalidTransition:
              value:
                type: "urn:ams:badge:invalid-transition"
                title: "Invalid badge state transition"
                status: 409
                detail: "Cannot transition badge from REVOKED to ACTIVE."
                instance: "/v1/badges/0197b0.../activate"
                traceId: "7c3f1a2b9d4e5f60"
...
    ValidationProblem:
      description: Semantic validation failure
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ValidationProblem' }
  schemas:
    Problem:
      type: object
      description: RFC 7807 problem details
      properties:
        type: { type: string, format: uri, default: "about:blank" }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string }
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items: { type: string }
    PageLinks:
      type: object
      properties:
        next: { type: [string, "null"], format: uri }
        prev: { type: [string, "null"], format: uri }
    PageMeta:
      type: object
      properties:
        pageSize: { type: integer }
...
    ReasonRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string, maxLength: 100 }
        comment: { type: string, maxLength: 1000 }
    BadgeIssueRequest:
      type: object
      required: [cardholderId, badgeType, validFrom, validUntil]
      properties:
        cardholderId: { type: string, format: uuid }
        badgeType: { type: string, enum: [RFID, QR, TEMPORARY_RFID] }
        validFrom: { type: string, format: date-time }
        validUntil: { type: string, format: date-time }
        siteId: { type: string, format: uuid, description: Issuing site (stock tracking) }
    Badge:
      type: object
      properties:
        badgeId: { type: string, format: uuid }
        cardholderId: { type: string, format: uuid }
        badgeType: { type: string, enum: [RFID, QR, TEMPORARY_RFID] }
        state: { type: string, enum: [REQUESTED, ISSUED, ACTIVE, SUSPENDED, REVOKED, EXPIRED, LOST, RETURNED] }
        validFrom: { type: string, format: date-time }
        validUntil: { type: string, format: date-time }
        issuedAt: { type: string, format: date-time }
        streamVersion: { type: integer, description: Aggregate version (optimistic concurrency) }
    BadgePage:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Badge' }
        links: { $ref: '#/components/schemas/PageLinks' }
        meta: { $ref: '#/components/schemas/PageMeta' }
...